new function symbol-amongl
o New function symbol-amongl.  Unlike amongl, the function symbol-amongl doesn't examine the car of an expression.

o In lenient-extended-realp, move the O(n) check to the end and convert the call of $freeof to symbol-amongl.  The function $freeof knows a more about meaning than does symbol-among.  So this is a change.